Refine training UI and modularize engine
