+++ src/App.jsx
@@ -1,369 +1,20 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 import { useEffect, useRef, useState } from 'react'
 
-const EPS = 1e-6
-const clamp = (x, lo, hi) => Math.min(Math.max(x, lo), hi)
-
-function sourceToF32RGB(src, targetMax = 512) {
-  const c = document.createElement('canvas')
-  const w0 = Math.max(1, src.width)
-  const h0 = Math.max(1, src.height)
-  const scale = Math.min(1, targetMax / Math.max(w0, h0))
-  const w = Math.max(1, Math.round(w0 * scale))
-  const h = Math.max(1, Math.round(h0 * scale))
-  c.width = w
-  c.height = h
-  const ctx = c.getContext('2d', { willReadFrequently: true })
-  if (!ctx) throw new Error('Canvas 2D context unavailable')
-  try {
-    ctx.drawImage(src, 0, 0, w, h)
-  } catch (err) {
-    console.error('Rasterization failed (drawImage)', err)
-    throw new Error('Rasterization failed (drawImage). The file may be corrupt or blocked.')
-  }
-  let imgData
-  try {
-    imgData = ctx.getImageData(0, 0, w, h)
-  } catch (err) {
-    console.error('Reading pixels failed (getImageData)', err)
-    throw new Error(
-      'Reading pixels failed (getImageData). If this is an SVG with external refs, re-export as PNG/JPG.',
-    )
-  }
-  const data = imgData.data
-  const f32 = new Float32Array(w * h * 3)
-  for (let i = 0; i < w * h; i++) {
-    f32[i * 3 + 0] = data[i * 4 + 0] / 255
-    f32[i * 3 + 1] = data[i * 4 + 1] / 255
-    f32[i * 3 + 2] = data[i * 4 + 2] / 255
-  }
-  return { data: f32, w, h }
-}
-
-async function fileToImageSource(file) {
-  if ('createImageBitmap' in window && typeof createImageBitmap === 'function') {
-    try {
-      const bmp = await createImageBitmap(file)
-      return { src: bmp, revoke: () => bmp.close?.() }
-    } catch (err) {
-      console.warn('createImageBitmap failed; falling back to HTMLImageElement', err)
-    }
-  }
-  const dataUrl = await new Promise((resolve, reject) => {
-    const fr = new FileReader()
-    fr.onerror = () => reject(new Error('FileReader failed'))
-    fr.onload = () => resolve(String(fr.result))
-    fr.readAsDataURL(file)
-  })
-  const img = await new Promise((resolve, reject) => {
-    const im = new Image()
-    im.crossOrigin = 'anonymous'
-    im.onload = () => resolve(im)
-    im.onerror = () => reject(new Error('Image decode failed'))
-    im.src = dataUrl
-  })
-  return { src: img, revoke: () => {} }
-}
-
-function sobelMagJS(rgb, w, h) {
-  const mag = new Float32Array(w * h)
-  const kx = [-1, 0, 1, -2, 0, 2, -1, 0, 1]
-  const ky = [-1, -2, -1, 0, 0, 0, 1, 2, 1]
-  for (let y = 0; y < h; y++) {
-    for (let x = 0; x < w; x++) {
-      let gx = 0
-      let gy = 0
-      for (let dy = -1; dy <= 1; dy++)
-        for (let dx = -1; dx <= 1; dx++) {
-          const xx = clamp(x + dx, 0, w - 1)
-          const yy = clamp(y + dy, 0, h - 1)
-          const i = yy * w + xx
-          const r = rgb[i * 3 + 0]
-          const g = rgb[i * 3 + 1]
-          const b = rgb[i * 3 + 2]
-          const gray = 0.299 * r + 0.587 * g + 0.114 * b
-          const kIdx = (dy + 1) * 3 + (dx + 1)
-          gx += gray * kx[kIdx]
-          gy += gray * ky[kIdx]
-        }
-      mag[y * w + x] = Math.sqrt(gx * gx + gy * gy)
-    }
-  }
-  return mag
-}
-
-function cpuTopKIndices(arr, K) {
-  const n = arr.length
-  const k = Math.min(K, n)
-  const idxs = new Array(k).fill(-1)
-  const vals = new Array(k).fill(-Infinity)
-  for (let i = 0; i < n; i++) {
-    const v = arr[i]
-    if (v > vals[k - 1]) {
-      let j = k - 1
-      while (j > 0 && v > vals[j - 1]) {
-        vals[j] = vals[j - 1]
-        idxs[j] = idxs[j - 1]
-        j--
-      }
-      vals[j] = v
-      idxs[j] = i
-    }
-  }
-  return idxs
-}
-
-function samplePositionsAndColorsJS(imgF32, w, h, prob2D, N) {
-  const prob = new Float32Array(prob2D)
-  let s = 0
-  for (let i = 0; i < prob.length; i++) s += prob[i]
-  s = s || 1
-  for (let i = 0; i < prob.length; i++) prob[i] /= s
-  const cdf = new Float32Array(prob.length)
-  let acc = 0
-  for (let i = 0; i < prob.length; i++) {
-    acc += prob[i]
-    cdf[i] = acc
-  }
-  const mu = new Float32Array(N * 2)
-  const colors = new Float32Array(N * 3)
-  for (let k = 0; k < N; k++) {
-    const r = Math.random()
-    let lo = 0
-    let hi = cdf.length - 1
-    while (lo < hi) {
-      const mid = (lo + hi) >> 1
-      if (cdf[mid] < r) lo = mid + 1
-      else hi = mid
-    }
-    const y = Math.floor(lo / w)
-    const x = lo % w
-    const idx = y * w + x
-    mu[k * 2 + 0] = x / (w - 1)
-    mu[k * 2 + 1] = y / (h - 1)
-    colors[k * 3 + 0] = imgF32[idx * 3 + 0]
-    colors[k * 3 + 1] = imgF32[idx * 3 + 1]
-    colors[k * 3 + 2] = imgF32[idx * 3 + 2]
-  }
-  return { mu, colors }
-}
-
-function psnrJS(a, b) {
-  let mse = 0
-  for (let i = 0; i < a.length; i++) {
-    const d = a[i] - b[i]
-    mse += d * d
-  }
-  mse /= a.length || 1
-  if (mse <= 0) return Infinity
-  return 10 * Math.log10(1 / (mse + EPS))
-}
-
-function float16Quantize(x) {
-  const f32 = x instanceof Float32Array ? x : new Float32Array(x)
-  const buf = new ArrayBuffer(f32.length * 2)
-  const dv = new DataView(buf)
-  for (let i = 0; i < f32.length; i++) dv.setUint16(i * 2, float32ToFloat16(f32[i]), true)
-  return new Uint8Array(buf)
-}
-
-function float32ToFloat16(val) {
-  const f32 = new Float32Array(1)
-  const i32 = new Int32Array(f32.buffer)
-  f32[0] = val
-  const x = i32[0]
-  const bits = (x >> 16) & 0x8000
-  const m = (x >> 12) & 0x07ff
-  const e = (x >> 23) & 0xff
-  if (e < 103) return bits
-  if (e > 142) return bits | 0x7c00
-  return bits | ((e - 112) << 10) | (m >> 1)
-}
-
-const WORKER_SRC = `
-  const EPS = 1e-6;
-  const clamp = (x, lo, hi) => Math.min(Math.max(x, lo), hi);
-  function buildTileBinsFromArrays(muA, sInvA, H, W, tileH, tileW) {
-    if (!muA || !sInvA) {
-      const tilesX = Math.ceil(W / tileW);
-      const tilesY = Math.ceil(H / tileH);
-      return { tilesX, tilesY, bins: Array.from({ length: tilesX * tilesY }, () => []) };
-    }
-    const tilesX = Math.ceil(W / tileW);
-    const tilesY = Math.ceil(H / tileH);
-    const bins = Array.from({ length: tilesX * tilesY }, () => []);
-    const N = muA.length / 2;
-    for (let i = 0; i < N; i++) {
-      const ux = muA[i * 2 + 0], uy = muA[i * 2 + 1];
-      const sxInv = sInvA[i * 2 + 0], syInv = sInvA[i * 2 + 1];
-      const cx = ux * (W - 1), cy = uy * (H - 1);
-      const rx = 3 * (1 / Math.max(sxInv, EPS)) * W;
-      const ry = 3 * (1 / Math.max(syInv, EPS)) * H;
-      const r = Math.max(rx, ry);
-      const x0 = Math.max(0, Math.floor((cx - r) / tileW));
-      const y0 = Math.max(0, Math.floor((cy - r) / tileH));
-      const x1 = Math.min(tilesX - 1, Math.floor((cx + r) / tileW));
-      const y1 = Math.min(tilesY - 1, Math.floor((cy + r) / tileH));
-      for (let ty = y0; ty <= y1; ty++) {
-        const row = ty * tilesX;
-        for (let tx = x0; tx <= x1; tx++) bins[row + tx].push(i);
-      }
-    }
-    return { tilesX, tilesY, bins };
-  }
-
-  let IMG = null;
-  let W = 0, H = 0;
-  let tileW = 32, tileH = 32;
-  let enableTiling = true;
-  let bins = null;
-
-  function computeStepStripe(mu, s_inv, theta, color, K, wantImage, doRebin, y0Stripe, y1Stripe, reqId) {
-    const N = mu.length / 2;
-    if (!IMG) throw new Error('Worker not initialized');
-    if (enableTiling && (doRebin || !bins)) bins = buildTileBinsFromArrays(mu, s_inv, H, W, tileH, tileW);
-
-    const out = wantImage ? new Float32Array((y1Stripe - y0Stripe + 1) * W * 3) : null;
-    const accW = new Float32Array(N);
-    const accTW = new Float32Array(N * 3);
-    const accX = new Float32Array(N);
-    const accY = new Float32Array(N);
-    const accXX = new Float32Array(N);
-    const accYY = new Float32Array(N);
-    const accXY = new Float32Array(N);
-
-    const tilesX = bins ? bins.tilesX : Math.ceil(W / tileW);
-    const tilesY = bins ? bins.tilesY : Math.ceil(H / tileH);
-
-    const ty0 = Math.max(0, Math.floor(y0Stripe / tileH));
-    const ty1 = Math.min(tilesY - 1, Math.floor(y1Stripe / tileH));
-
-    for (let ty = ty0; ty <= ty1; ty++) {
-      const xTiles = tilesX;
-      const y0 = Math.max(y0Stripe, ty * tileH);
-      const y1 = Math.min(y1Stripe + 1, (ty + 1) * tileH);
-      for (let tx = 0; tx < xTiles; tx++) {
-        const x0 = tx * tileW;
-        const x1 = Math.min(W, x0 + tileW);
-        const list = enableTiling && bins ? bins.bins[ty * tilesX + tx] : null;
-        for (let y = y0; y < y1; y++) {
-          for (let x = x0; x < x1; x++) {
-            const cx = x / (W - 1), cy = y / (H - 1);
-            const topIdx = [];
-            const topVal = [];
-            const scan = (i) => {
-              const dx = cx - mu[i * 2 + 0];
-              const dy = cy - mu[i * 2 + 1];
-              const th = theta[i];
-              const c = Math.cos(th), s = Math.sin(th);
-              const dxp = c * dx + s * dy;
-              const dyp = -s * dx + c * dy;
-              const sx = s_inv[i * 2 + 0], sy = s_inv[i * 2 + 1];
-              const z = (dxp * sx) * (dxp * sx) + (dyp * sy) * (dyp * sy);
-              const g = Math.exp(-0.5 * z);
-              if (topIdx.length < K) {
-                let j = topIdx.length;
-                topIdx.push(i);
-                topVal.push(g);
-                while (j > 0 && topVal[j] > topVal[j - 1]) {
-                  const ti = topIdx[j - 1];
-                  topIdx[j - 1] = topIdx[j];
-                  topIdx[j] = ti;
-                  const tv = topVal[j - 1];
-                  topVal[j - 1] = topVal[j];
-                  topVal[j] = tv;
-                  j--;
-                }
-              } else if (g > topVal[topVal.length - 1]) {
-                topIdx[topIdx.length - 1] = i;
-                topVal[topVal.length - 1] = g;
-                let j = topIdx.length - 1;
-                while (j > 0 && topVal[j] > topVal[j - 1]) {
-                  const ti = topIdx[j - 1];
-                  topIdx[j - 1] = topIdx[j];
-                  topIdx[j] = ti;
-                  const tv = topVal[j - 1];
-                  topVal[j - 1] = topVal[j];
-                  topVal[j] = tv;
-                  j--;
-                }
-              }
-            };
-            if (list && list.length) {
-              for (let ii = 0; ii < list.length; ii++) scan(list[ii]);
-            } else {
-              for (let i = 0; i < N; i++) scan(i);
-            }
-            let wsum = EPS;
-            for (let k = 0; k < topVal.length; k++) wsum += topVal[k];
-            const baseGlobal = (y * W + x) * 3;
-            let r = 0, g = 0, b = 0;
-            for (let k = 0; k < topIdx.length; k++) {
-              const wi = topVal[k] / wsum;
-              const gi = topIdx[k];
-              r += wi * color[gi * 3 + 0];
-              g += wi * color[gi * 3 + 1];
-              b += wi * color[gi * 3 + 2];
-              accW[gi] += wi;
-              accTW[gi * 3 + 0] += wi * IMG[baseGlobal + 0];
-              accTW[gi * 3 + 1] += wi * IMG[baseGlobal + 1];
-              accTW[gi * 3 + 2] += wi * IMG[baseGlobal + 2];
-              const cxn = cx, cyn = cy;
-              accX[gi] += wi * cxn;
-              accY[gi] += wi * cyn;
-              accXX[gi] += wi * cxn * cxn;
-              accYY[gi] += wi * cyn * cyn;
-              accXY[gi] += wi * cxn * cyn;
-            }
-            if (out) {
-              const row = y - y0Stripe;
-              const baseLocal = (row * W + x) * 3;
-              out[baseLocal + 0] = r;
-              out[baseLocal + 1] = g;
-              out[baseLocal + 2] = b;
-            }
-          }
-        }
-      }
-    }
-
-    const msg = { type: 'stepResult', reqId, accW, accTW, accX, accY, accXX, accYY, accXY, y0: y0Stripe, y1: y1Stripe };
-    const transfers = [accW.buffer, accTW.buffer, accX.buffer, accY.buffer, accXX.buffer, accYY.buffer, accXY.buffer];
-    if (out) {
-      msg.out = out;
-      transfers.push(out.buffer);
-    }
-    postMessage(msg, transfers);
-  }
-
-  onmessage = (ev) => {
-    const m = ev.data;
-    if (m.type === 'init') {
-      IMG = m.img;
-      W = m.W;
-      H = m.H;
-      tileW = m.tileW;
-      tileH = m.tileH;
-      enableTiling = !!m.enableTiling;
-      bins = null;
-      postMessage({ type: 'inited' });
-    } else if (m.type === 'step') {
-      computeStepStripe(m.mu, m.s_inv, m.theta, m.color, m.K, !!m.wantImage, !!m.doRebin, m.y0, m.y1, m.reqId);
-    } else if (m.type === 'setTiling') {
-      tileW = m.tileW;
-      tileH = m.tileH;
-      enableTiling = !!m.enableTiling;
-      bins = null;
-    }
-  };
-`
-
-function makeWorker() {
-  const blob = new Blob([WORKER_SRC], { type: 'text/javascript' })
-  const url = URL.createObjectURL(blob)
-  const w = new Worker(url, { type: 'classic' })
-  return { worker: w, url }
-}
+import InfoPopover from './components/InfoPopover'
+import { fileToImageSource, sourceToF32RGB } from './lib/imageLoader'
+import { clamp, psnrJS } from './lib/gaussianMath'
+import { float16Quantize } from './lib/quantization'
+import {
+  composeOutput,
+  computeErrorField,
+  extendModelWithErrors,
+  initializeParameters,
+  mergeAccumulators,
+  selectErrorIndices,
+  updateParametersFromAccumulators,
+} from './lib/training'
+import { makeWorker } from './lib/workerFactory'
 
 const buttonBase =
   'inline-flex items-center justify-center gap-2 rounded-full px-4 py-2 text-sm font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 focus-visible:ring-sky-400 focus-visible:ring-offset-slate-950 disabled:cursor-not-allowed disabled:opacity-60'
@@ -416,81 +67,6 @@ const statusStyles = {
   },
 }
 
-function Diagnostics() {
-  const [diag, setDiag] = useState('idle')
-  const run = async () => {
-    try {
-      setDiag('running')
-      const { worker, url } = makeWorker()
-      const W = 8
-      const H = 8
-      const img = new Float32Array(W * H * 3)
-      for (let y = 0; y < H; y++)
-        for (let x = 0; x < W; x++) {
-          const i = (y * W + x) * 3
-          img[i] = x / (W - 1)
-          img[i + 1] = y / (H - 1)
-          img[i + 2] = 0.5
-        }
-      const waitType = (type) =>
-        new Promise((res) => {
-          const handler = (ev) => {
-            if (ev.data?.type === type) {
-              worker.removeEventListener('message', handler)
-              res(ev.data)
-            }
-          }
-          worker.addEventListener('message', handler)
-        })
-      worker.postMessage({ type: 'init', img, W, H, tileW: 4, tileH: 4, enableTiling: true }, [img.buffer])
-      await waitType('inited')
-      const mu = new Float32Array([0.25, 0.25, 0.75, 0.75])
-      const s_inv = new Float32Array([W - 1, H - 1, W - 1, H - 1])
-      const theta = new Float32Array([0, 0])
-      const color = new Float32Array([1, 0, 0, 0, 1, 0])
-      const reqId = 1234
-      const got = new Promise((res) => {
-        const handler = (ev) => {
-          const d = ev.data
-          if (d?.type === 'stepResult' && d.reqId === reqId) {
-            worker.removeEventListener('message', handler)
-            res(d)
-          }
-        }
-        worker.addEventListener('message', handler)
-      })
-      worker.postMessage({
-        type: 'step',
-        reqId,
-        mu,
-        s_inv,
-        theta,
-        color,
-        K: 2,
-        wantImage: true,
-        doRebin: true,
-        y0: 0,
-        y1: H - 1,
-      })
-      const d = await got
-      URL.revokeObjectURL(url)
-      worker.terminate()
-      if (!d.out || d.out.length !== W * H * 3) throw new Error('render size mismatch')
-      setDiag('ok')
-    } catch (e) {
-      setDiag('fail: ' + (e?.message || e))
-    }
-  }
-  return (
-    <div className="flex items-center gap-3 text-slate-300">
-      <button className={ghostButton} onClick={run}>
-        Run diagnostics
-      </button>
-      <span className="text-xs font-medium uppercase tracking-wide text-slate-400">{diag}</span>
-    </div>
-  )
-}
-
 export default function ImageGSApp() {
   const [imgJS, setImgJS] = useState(null)
   const [imgSize, setImgSize] = useState({ w: 0, h: 0 })
@@ -502,6 +78,7 @@ export default function ImageGSApp() {
   const [lambdaInit, setLambdaInit] = useState(0.3)
   const [steps, setSteps] = useState(1500)
   const [stepDelayMs, setStepDelayMs] = useState(0)
+  const [addEvery, setAddEvery] = useState(400)
 
   const [lrColor, setLrColor] = useState(0.4)
   const [lrMu, setLrMu] = useState(0.25)
@@ -680,27 +257,9 @@ export default function ImageGSApp() {
 
   async function initParamsJS() {
     if (!imgJS) return
-    const { w, h, data } = imgJS
-    const grad = sobelMagJS(data, w, h)
-    let sum = 0
-    for (let i = 0; i < grad.length; i++) sum += grad[i]
-    const uniform = 1 / (w * h)
-    const probs = new Float32Array(w * h)
-    const lmb = lambdaInit
-    for (let i = 0; i < w * h; i++) probs[i] = (1 - lmb) * (grad[i] / (sum || 1)) + lmb * uniform
-
-    const Ng0 = Math.max(1, Math.floor(budget / 2))
-    const { mu, colors } = samplePositionsAndColorsJS(data, w, h, probs, Ng0)
-    const s_inv = new Float32Array(Ng0 * 2)
-    for (let i = 0; i < Ng0; i++) {
-      s_inv[i * 2 + 0] = (w - 1) / 5
-      s_inv[i * 2 + 1] = (h - 1) / 5
-    }
-    const theta = new Float32Array(Ng0)
-    for (let i = 0; i < Ng0; i++) theta[i] = Math.random() * Math.PI
-
-    varsRef.current = { mu, s_inv, theta, color: colors }
-    curNgRef.current = Ng0
+    const { vars, count } = initializeParameters(imgJS, budget, lambdaInit)
+    varsRef.current = vars
+    curNgRef.current = count
     broadcastSetTiling()
   }
 
@@ -713,46 +272,6 @@ export default function ImageGSApp() {
       theta: new Float32Array(v.theta),
       color: new Float32Array(v.color),
     }
-  }
-
-  function mergeAccumulators(parts, N) {
-    const accW = new Float32Array(N)
-    const accTW = new Float32Array(N * 3)
-    const accX = new Float32Array(N)
-    const accY = new Float32Array(N)
-    const accXX = new Float32Array(N)
-    const accYY = new Float32Array(N)
-    const accXY = new Float32Array(N)
-    for (const p of parts) {
-      const aW = p.accW
-      const aT = p.accTW
-      const aX = p.accX
-      const aY = p.accY
-      const aXX = p.accXX
-      const aYY = p.accYY
-      const aXY = p.accXY
-      for (let i = 0; i < N; i++) accW[i] += aW[i]
-      for (let i = 0; i < N * 3; i++) accTW[i] += aT[i]
-      for (let i = 0; i < N; i++) {
-        accX[i] += aX[i]
-        accY[i] += aY[i]
-        accXX[i] += aXX[i]
-        accYY[i] += aYY[i]
-        accXY[i] += aXY[i]
-      }
-    }
-    return { accW, accTW, accX, accY, accXX, accYY, accXY }
-  }
-
-  function composeOut(parts, W, H) {
-    const out = new Float32Array(W * H * 3)
-    for (const p of parts) {
-      if (!p.out) continue
-      const y0 = p.y0
-      const stripe = p.out
-      out.set(stripe, y0 * W * 3)
-    }
-    return out
   }
 
   async function train() {
@@ -770,75 +289,14 @@ export default function ImageGSApp() {
     setStatus('done')
   }
 
-  function updateParamsFromAcc(N, pkg) {
-    const v = varsRef.current
-    if (!v) return
-    for (let i = 0; i < N; i++) {
-      const wsum = pkg.accW[i] + 1e-6
-      const tR = pkg.accTW[i * 3 + 0] / wsum
-      const tG = pkg.accTW[i * 3 + 1] / wsum
-      const tB = pkg.accTW[i * 3 + 2] / wsum
-      v.color[i * 3 + 0] = (1 - lrColor) * v.color[i * 3 + 0] + lrColor * tR
-      v.color[i * 3 + 1] = (1 - lrColor) * v.color[i * 3 + 1] + lrColor * tG
-      v.color[i * 3 + 2] = (1 - lrColor) * v.color[i * 3 + 2] + lrColor * tB
-    }
-    const w = imgSize.w
-    const h = imgSize.h
-    const stdMinX = 1 / Math.max(w - 1, 1)
-    const stdMinY = 1 / Math.max(h - 1, 1)
-    for (let i = 0; i < N; i++) {
-      const wsum = pkg.accW[i]
-      if (wsum < 1e-6) continue
-      const mx = pkg.accX[i] / wsum
-      const my = pkg.accY[i] / wsum
-      const muX = v.mu[i * 2 + 0]
-      const muY = v.mu[i * 2 + 1]
-      v.mu[i * 2 + 0] = (1 - lrMu) * muX + lrMu * mx
-      v.mu[i * 2 + 1] = (1 - lrMu) * muY + lrMu * my
-      const cxx = Math.max(0, pkg.accXX[i] / wsum - mx * mx)
-      const cyy = Math.max(0, pkg.accYY[i] / wsum - my * my)
-      const cxy = pkg.accXY[i] / wsum - mx * my
-      const tr = cxx + cyy
-      const det = cxx * cyy - cxy * cxy
-      const disc = Math.max(0, tr * tr - 4 * det)
-      const s = Math.sqrt(disc)
-      const l1 = 0.5 * (tr + s)
-      const l2 = 0.5 * (tr - s)
-      let vx
-      let vy
-      if (Math.abs(cxy) > 1e-12) {
-        vx = l1 - cyy
-        vy = cxy
-      } else if (cxx >= cyy) {
-        vx = 1
-        vy = 0
-      } else {
-        vx = 0
-        vy = 1
-      }
-      const n = Math.hypot(vx, vy) || 1
-      vx /= n
-      vy /= n
-      let th0 = v.theta[i] % Math.PI
-      if (th0 < 0) th0 += Math.PI
-      let thT = Math.atan2(vy, vx)
-      thT %= Math.PI
-      if (thT < 0) thT += Math.PI
-      if (Math.abs(thT - th0) > Math.PI / 2) {
-        if (thT > th0) thT -= Math.PI
-        else thT += Math.PI
-      }
-      const thNew = th0 + (thT - th0) * lrShape
-      v.theta[i] = ((thNew % Math.PI) + Math.PI) % Math.PI
-      const std1 = Math.sqrt(Math.max(l1, 0))
-      const std2 = Math.sqrt(Math.max(l2, 0))
-      const sxInvTarget = 1 / Math.max(std1, Math.min(stdMinX, stdMinY))
-      const syInvTarget = 1 / Math.max(std2, Math.min(stdMinX, stdMinY))
-      const sx0 = v.s_inv[i * 2 + 0]
-      const sy0 = v.s_inv[i * 2 + 1]
-      v.s_inv[i * 2 + 0] = clamp((1 - lrShape) * sx0 + lrShape * sxInvTarget, 0.1, (w - 1) * 4)
-      v.s_inv[i * 2 + 1] = clamp((1 - lrShape) * sy0 + lrShape * syInvTarget, 0.1, (h - 1) * 4)
-    }
+  function applyAccumulatorUpdate(pkg) {
+    if (!varsRef.current) return
+    updateParametersFromAccumulators(varsRef.current, pkg, {
+      lrColor,
+      lrMu,
+      lrShape,
+      size: imgSize,
+    })
   }
 
   async function renderFrameViaPool(doRebin = false) {
@@ -859,16 +317,16 @@ export default function ImageGSApp() {
       promises.push(postStepToWorker(i, payload))
     }
     const results = await Promise.all(promises)
-    return composeOut(results, W, H)
+    return composeOutput(results, W, H)
   }
 
   async function trainJS_pool() {
     const pool = poolRef.current
     if (!pool.workers.length) return
     const Ntotal = budget
-    const addEvery = 400
     const addChunk = Math.max(1, Math.floor(Ntotal / 8))
     const { w: W, h: H } = imgSize
+    const addFrequency = addEvery > 0 ? addEvery : null
 
     for (let step = 1; step <= steps; step++) {
       if (stopRef.current) break
@@ -890,10 +348,10 @@ export default function ImageGSApp() {
       }
       const results = await Promise.all(promises)
       const merged = mergeAccumulators(results, N)
-      updateParamsFromAcc(N, merged)
+      applyAccumulatorUpdate(merged)
 
       if (wantImage) {
-        const out = composeOut(results, W, H)
+        const out = composeOutput(results, W, H)
         drawOnCanvasF32RGB(out, W, H)
         lastDrawRef.current = performance.now()
         const ps = psnrJS(out, imgJS.data)
@@ -902,7 +360,7 @@ export default function ImageGSApp() {
         setMetrics((m) => ({ ...m, step, n: curNgRef.current, pool: stripes }))
       }
 
-      if (step % addEvery === 0 && curNgRef.current < Ntotal) {
+      if (addFrequency && step % addFrequency === 0 && curNgRef.current < Ntotal) {
         await addGaussiansByErrorJS(Math.min(addChunk, Ntotal - curNgRef.current))
       }
       if (stepDelayMs > 0) await new Promise((r) => setTimeout(r, stepDelayMs))
@@ -914,49 +372,16 @@ export default function ImageGSApp() {
 
   async function addGaussiansByErrorJS(nNew) {
     const out = await renderFrameViaPool(false)
-    if (!out) return
+    if (!out || !imgJS) return
     const { data: tgt, w: W, h: H } = imgJS
-    const err = new Float32Array(W * H)
-    for (let i = 0; i < W * H; i++) {
-      err[i] =
-        (Math.abs(out[i * 3 + 0] - tgt[i * 3 + 0]) +
-          Math.abs(out[i * 3 + 1] - tgt[i * 3 + 1]) +
-          Math.abs(out[i * 3 + 2] - tgt[i * 3 + 2])) /
-        3
-    }
-    const idxs = cpuTopKIndices(err, Math.min(nNew, W * H))
+    const err = computeErrorField(out, tgt, W, H)
+    const idxs = selectErrorIndices(err, Math.min(nNew, W * H))
 
     const v = varsRef.current
     if (!v) return
-    const NgOld = v.mu.length / 2
-    const NgNew = NgOld + nNew
-    const mu2 = new Float32Array(NgNew * 2)
-    mu2.set(v.mu)
-    const s2 = new Float32Array(NgNew * 2)
-    s2.set(v.s_inv)
-    const th2 = new Float32Array(NgNew)
-    th2.set(v.theta)
-    const c2 = new Float32Array(NgNew * 3)
-    c2.set(v.color)
-
-    for (let j = 0; j < idxs.length; j++) {
-      const id = idxs[j]
-      const y = Math.floor(id / W)
-      const x = id % W
-      const base = y * W + x
-      const i = NgOld + j
-      mu2[i * 2 + 0] = x / (W - 1)
-      mu2[i * 2 + 1] = y / (H - 1)
-      s2[i * 2 + 0] = (W - 1) / 5
-      s2[i * 2 + 1] = (H - 1) / 5
-      th2[i] = Math.random() * Math.PI
-      c2[i * 3 + 0] = tgt[base * 3 + 0]
-      c2[i * 3 + 1] = tgt[base * 3 + 1]
-      c2[i * 3 + 2] = tgt[base * 3 + 2]
-    }
-
-    varsRef.current = { mu: mu2, s_inv: s2, theta: th2, color: c2 }
-    curNgRef.current = NgNew
+    const { vars, count } = extendModelWithErrors(v, idxs, imgJS)
+    varsRef.current = vars
+    curNgRef.current = count
     broadcastSetTiling()
   }
 
@@ -1020,13 +445,13 @@ export default function ImageGSApp() {
           <div className="flex flex-col gap-6 lg:flex-row lg:items-start lg:justify-between">
             <div className="space-y-4">
               <span className="inline-flex w-fit items-center gap-2 rounded-full border border-white/10 bg-white/10 px-3 py-1 text-xs font-semibold uppercase tracking-[0.3em] text-slate-300/80">
-                Portfolio build
+                Client showcase
               </span>
               <h1 className="text-4xl font-semibold tracking-tight text-white sm:text-5xl">
                 Image-GS — 2D Gaussian Splatting in the Browser
               </h1>
               <p className="max-w-2xl text-base text-slate-300">
-                Upload an image and watch a pool of web workers optimize a Gaussian splat representation in real time. The entire experience runs on pure JavaScript – no WebGL, WASM, or server renderers required.
+                Upload an image to see a refined Gaussian splat reinterpretation bloom in real time. The entire pipeline runs in pure JavaScript—no WebGL, WASM, or servers required—making it ideal for polished, client-facing showcases.
               </p>
             </div>
             <span className={`${statusBadgeBase} ${statusMeta.classes}`}>
@@ -1035,7 +460,7 @@ export default function ImageGSApp() {
             </span>
           </div>
           <p className="max-w-3xl text-sm text-slate-400">
-            Drag in a photo, tune the parameters, and the renderer will progressively refine the preview. This build is CSP-safe, uses dedicated web workers for EM updates, and automatically downsamples sources to 512px for responsive demos.
+            Tailor the experience with production-ready controls. Dedicated worker orchestration keeps playback responsive, while smart 512&nbsp;px preprocessing and CSP-friendly code ensure effortless deployment to high-security environments.
           </p>
         </header>
 
@@ -1078,10 +503,9 @@ export default function ImageGSApp() {
                   >
                     Save .igs
                   </button>
-                  <Diagnostics />
                 </div>
                 <p className="text-sm text-slate-400">
-                  Training spawns a worker pool sized to your CPU. Capture a snapshot at any time or export the learned splats as a compact <code className="rounded bg-white/10 px-1 py-0.5 text-xs">.igs</code> file.
+                  Training orchestrates a worker pool tuned to each visitor’s hardware, delivering smooth, reliable updates. Capture snapshots at any moment or export the learned splats as a compact <code className="rounded bg-white/10 px-1 py-0.5 text-xs">.igs</code> file.
                 </p>
               </div>
             </div>
@@ -1102,7 +526,13 @@ export default function ImageGSApp() {
                     </h3>
                     <div className="mt-4 grid gap-4 sm:grid-cols-2 xl:grid-cols-3">
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">K</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">K</span>
+                          <InfoPopover
+                            title="Nearest splats (K)"
+                            description="Controls how many Gaussians blend together for each pixel. Higher values can sharpen detail at the cost of more computation."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={K}
@@ -1113,7 +543,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Budget (N)</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Budget (N)</span>
+                          <InfoPopover
+                            title="Gaussian budget"
+                            description="Sets the maximum number of splats the optimizer can allocate. Larger budgets capture more nuance but require more memory and processing."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={budget}
@@ -1124,7 +560,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">λ_init</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">λ_init</span>
+                          <InfoPopover
+                            title="Initialization blend"
+                            description="Balances gradient-based seeding with a uniform distribution. Higher values produce broader coverage before training begins."
+                          />
+                        </div>
                         <input
                           type="number"
                           step="0.05"
@@ -1136,7 +578,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Steps</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Steps</span>
+                          <InfoPopover
+                            title="Optimization steps"
+                            description="Defines how many EM iterations run during training. More steps yield higher fidelity until convergence."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={steps}
@@ -1147,13 +595,39 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Delay (ms)</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Delay (ms)</span>
+                          <InfoPopover
+                            title="Step delay"
+                            description="Adds a pause between iterations in milliseconds. Useful for showcasing the progression at a relaxed pace."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={stepDelayMs}
                           min={0}
                           max={1000}
                           onChange={(e) => setStepDelayMs(parseInt(e.target.value, 10) || 0)}
+                          className={inputClass}
+                        />
+                      </label>
+                      <label className="flex flex-col gap-2 text-sm text-slate-300">
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Add splats every</span>
+                          <InfoPopover
+                            title="Auto-growth cadence"
+                            description="Determines how often new Gaussians are introduced based on reconstruction error. Set to 0 to keep the count fixed."
+                          />
+                        </div>
+                        <input
+                          type="number"
+                          value={addEvery}
+                          min={0}
+                          max={5000}
+                          onChange={(e) => {
+                            const v = parseInt(e.target.value, 10)
+                            setAddEvery(Number.isFinite(v) ? Math.max(0, v) : 0)
+                          }}
                           className={inputClass}
                         />
                       </label>
@@ -1166,7 +640,13 @@ export default function ImageGSApp() {
                     </h3>
                     <div className="mt-4 grid gap-4 sm:grid-cols-2 xl:grid-cols-3">
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Pool size</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Pool size</span>
+                          <InfoPopover
+                            title="Worker pool"
+                            description="Specifies how many web workers optimize the image in parallel. Match this to the number of CPU cores available for smoother playback."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={poolSize}
@@ -1177,7 +657,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300 sm:col-span-2 xl:col-span-1">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tiling</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tiling</span>
+                          <InfoPopover
+                            title="Adaptive tiling"
+                            description="Caches Gaussians into spatial bins so each worker touches fewer splats per pixel. Disable to profile the raw solver."
+                          />
+                        </div>
                         <div className="flex items-center gap-3 rounded-xl border border-white/10 bg-slate-950/50 px-4 py-3">
                           <input
                             type="checkbox"
@@ -1194,7 +680,13 @@ export default function ImageGSApp() {
                         </div>
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tile W</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tile W</span>
+                          <InfoPopover
+                            title="Tile width"
+                            description="Width of each tiling bucket in pixels. Smaller tiles improve accuracy but require more bin rebuilds."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={tileW}
@@ -1209,7 +701,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tile H</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Tile H</span>
+                          <InfoPopover
+                            title="Tile height"
+                            description="Height of each tiling bucket in pixels. Tune alongside tile width for the desired performance profile."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={tileH}
@@ -1224,7 +722,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Rebin every</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">Rebin every</span>
+                          <InfoPopover
+                            title="Rebin cadence"
+                            description="How frequently (in steps) the tiling bins are rebuilt. Set to 0 to reuse bins indefinitely."
+                          />
+                        </div>
                         <input
                           type="number"
                           value={rebinEvery}
@@ -1243,7 +747,13 @@ export default function ImageGSApp() {
                     </h3>
                     <div className="mt-4 grid gap-4 sm:grid-cols-2 xl:grid-cols-3">
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR Color</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR Color</span>
+                          <InfoPopover
+                            title="Color learning rate"
+                            description="Controls how quickly Gaussian colors adapt toward the target image. Lower values smooth the transition; higher values react faster."
+                          />
+                        </div>
                         <input
                           type="number"
                           step="0.05"
@@ -1255,7 +765,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR μ</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR μ</span>
+                          <InfoPopover
+                            title="Position learning rate"
+                            description="Adjusts how rapidly Gaussian centers move toward the weighted average of the pixels they explain."
+                          />
+                        </div>
                         <input
                           type="number"
                           step="0.05"
@@ -1267,7 +783,13 @@ export default function ImageGSApp() {
                         />
                       </label>
                       <label className="flex flex-col gap-2 text-sm text-slate-300">
-                        <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR shape</span>
+                        <div className="flex items-center justify-between">
+                          <span className="text-xs font-semibold uppercase tracking-wide text-slate-400">LR shape</span>
+                          <InfoPopover
+                            title="Shape learning rate"
+                            description="Controls the update speed for Gaussian covariance and orientation. Lower settings stabilize anisotropic splats."
+                          />
+                        </div>
                         <input
                           type="number"
                           step="0.05"
@@ -1329,31 +851,31 @@ export default function ImageGSApp() {
           <aside className="space-y-8">
             <div className={glassCardClass}>
               <div className="prose prose-invert max-w-none">
-                <h2>Under the hood</h2>
+                <h2>Technology highlights</h2>
                 <p>
-                  Each optimization step fans out across a dedicated worker pool sized to your hardware. Workers operate on row stripes, return partial accumulators, and the main thread simply reduces and composites the frame at a steady cadence.
+                  Each optimization step fans out across a dedicated worker pool sized to the visitor’s hardware. Workers solve their row stripes in parallel, stream back accumulators, and the main thread composites a fresh frame with cinema-grade cadence.
                 </p>
                 <h3>Adaptive tiling</h3>
                 <p>
-                  Keep tiling enabled to maintain roughly O(K) work per pixel. Rebin schedules rebuild each worker&apos;s tile→Gaussian bins whenever μ, θ, or σ⁻¹ drift, keeping hot spots responsive without reloading the model.
+                  Gaussian bins are rebuilt on schedule so hotspots stay responsive without reloading the model. The result is near O(K) complexity per pixel even as splats reshape over time.
                 </p>
                 <h3>Preview sizing</h3>
                 <p>
-                  Inputs are automatically scaled so the longest edge is 512&nbsp;px. It keeps demos smooth while preserving enough detail for crisp exports and portfolio-ready screen captures.
+                  Sources are automatically scaled to a 512&nbsp;px long edge, preserving crisp detail while ensuring buttery-smooth playback on modern laptops and tablets.
                 </p>
-                <h3>Next steps</h3>
+                <h3>Feature set</h3>
                 <ul>
-                  <li>Increase the budget and lower learning rates for photorealistic reconstructions.</li>
-                  <li>Toggle tiling off to profile the difference in worker throughput.</li>
-                  <li>Export a <code>.igs</code> file and embed the splats in another project.</li>
+                  <li>High budgets with tempered learning rates unlock photorealistic reconstructions.</li>
+                  <li>Tiling controls expose clear performance levers for live demos and deep dives.</li>
+                  <li>Export the learned <code>.igs</code> file to integrate splats into wider client experiences.</li>
                 </ul>
               </div>
             </div>
 
             <div className="rounded-3xl border border-sky-500/30 bg-sky-500/10 p-6 shadow-xl shadow-sky-900/40 backdrop-blur">
-              <h3 className="text-sm font-semibold uppercase tracking-[0.3em] text-slate-100">Tip</h3>
+              <h3 className="text-sm font-semibold uppercase tracking-[0.3em] text-slate-100">Spotlight</h3>
               <p className="mt-3 text-sm text-slate-100/80">
-                Square images converge fastest, but any aspect ratio works. Pair this viewer with a static site host and you&apos;ve got a performant, CSP-friendly Gaussian splat demo ready for your portfolio.
+                Square imagery converges in record time, yet any aspect ratio renders gracefully. Host the viewer on a static site to deliver an elegant, CSP-friendly Gaussian splat experience to clients worldwide.
               </p>
             </div>
           </aside>
